backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Config for uv"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asgiref==3.8.1",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "charset-normalizer==3.4.1",
    "cryptography==44.0.2",
    "defusedxml==0.7.1",
    "django==5.1.6",
    "django-debug-toolbar==5.0.1",
    "django-extensions==3.2.3",
    "django-filter==25.1",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.5.0",
    "djoser==2.3.1",
    "flake8==7.1.2",
    "gunicorn==23.0.0",
    "idna==3.10",
    "mccabe==0.7.0",
    "oauthlib==3.2.2",
    "packaging==24.2",
    "pillow==11.1.0",
    "psycopg2-binary==2.9.10",
    "pycodestyle==2.12.1",
    "pycparser==2.22",
    "pyflakes==3.2.0",
    "pyjwt==2.9.0",
    "python-dotenv==1.0.1",
    "python3-openid==3.2.0",
    "reportlab==4.3.1",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "ruff==0.8.0",
    "social-auth-app-django==5.4.3",
    "social-auth-core==4.5.6",
    "sqlparse==0.5.3",
    "typing-extensions==4.12.2",
    "urllib3==2.3.0",
]
